refactor(migrations): support getting statistics in tsurge testing
This allows us to test statistics in future commits.
